Screen Splashes
The fade in and Fade out should work. When I run the application it
fades in and out naturally, while other times it does not fade in and
out naturally, rather it jumps from one intent to the next.

I will most likely ask for help from another group to seek their
suggestions.

Let me know if you are getting the fade in and fade out effect