Changes are in Log In , and Vocab Game
Vocab Game is nearly complete. All fields are filled out. You will
notice.  However, there is still much to implement if i would like to
make it more complex

The Splashes now fade into the Log In page, instead of the about us
page.

The log in page has a link word that takes the user to the about us
page.

the log in page also now has the a register button.
The toolbar has been modified.